Adding a scaffold generator for the Admin plugin.
Now you can generate scaffolds that match the admin look and feel and conventions.

This scaffolds are ready for user friendly permissions (ie not showing links that user is not allowed to press).

In order to generate admin scaffolds you just need to replace 

{{{
./script/generate scaffold
}}} 

with

{{{
./script/generate admin_scaffold
}}} 

It is not fully tested yet, but it's a good start.


git-svn-id: http://svn.akelos.org@524 a2fa5c27-f921-0410-a72c-bf682d381be0