Controller considers ->respondTo() (content_type) when picking a template.
Controller picks index.html or index.xml as the template name (instead of plain 'index'). For compatibility it is allowed to omit the .html extension on your actual template-files. So you can name your file index.tpl or index.html.tpl. The latter one is preferred, the other one should be deprecated.


git-svn-id: http://svn.akelos.org@1020 a2fa5c27-f921-0410-a72c-bf682d381be0