Adding basic port of rake build tool to PHP.
It only supports a very basic set of rake's features. If you miss the handy FileList it can be easily emulated using PHP's glob()

At the end of lib/makelos.php you'll see the list of tasks that will be bundled into this plugin. Feel free to post your tasks on the Trac site or the Wiki. 


git-svn-id: http://svn.akelos.org@1413 a2fa5c27-f921-0410-a72c-bf682d381be0